LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-csp.https-expected.txt

CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to https://www1.web-platform.test:9443/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE&fetch3 because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to https://www1.web-platform.test:9443/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE&fetch4 because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Blocked by Content Security Policy.
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE&fetch4 due to access control checks.

PASS Verify CSP control of fetch() in a Service Worker
